move WKWebKit customization to DEVWKWebKit
