+++ DEV-Simple/views/ViewController.swift
@@ -25,39 +25,7 @@ class ViewController: UIViewController {
 
     @IBOutlet weak var backButton: UIBarButtonItem!
     @IBOutlet weak var forwardButton: UIBarButtonItem!
-    @IBOutlet lazy var webView: WKWebView! = {
-
-        if !UIAccessibility.isInvertColorsEnabled {
-            return WKWebView()
-        }
-
-        guard let path = Bundle.main.path(forResource: "invertedImages", ofType: "css") else {
-            return WKWebView()
-        }
-
-        let cssString = try? String(contentsOfFile: path).components(separatedBy: .newlines).joined()
-        let source = """
-        var style = document.createElement('style');
-        style.innerHTML = '\(cssString)';
-        document.head.appendChild(style);
-        """
-
-        let userScript = WKUserScript(source: source,
-                                      injectionTime: .atDocumentEnd,
-                                      forMainFrameOnly: true)
-
-        let userContentController = WKUserContentController()
-        userContentController.addUserScript(userScript)
-
-        let configuration = WKWebViewConfiguration()
-        configuration.userContentController = userContentController
-
-        let webView = WKWebView(frame: .zero,
-                                configuration: configuration)
-
-        webView.accessibilityIgnoresInvertColors = true
-        return webView
-    }()
+    @IBOutlet weak var webView: DEVWKWebView!
 
     @IBOutlet weak var safariButton: UIBarButtonItem!
     @IBOutlet weak var activityIndicator: UIActivityIndicatorView!
@@ -83,7 +51,6 @@ class ViewController: UIViewController {
         backButton.isEnabled = false
         forwardButton.isEnabled = false
         webView.navigationDelegate = self
-        webView.customUserAgent = "DEV-Native-ios"
         webView.scrollView.scrollIndicatorInsets.top = view.safeAreaInsets.top + 50
         webView.load(devToURL)
         webView.configuration.allowsInlineMediaPlayback = true
